Fbx generator refactored
