+++ Code/Source/URDF/Fbx.h
@@ -9,94 +9,70 @@
 #pragma once
 
 #include <string>
-#include <any>
+
+#include "FbxNode.h"
 
 namespace ROS2
 {
-    ////////////////////////////////////////////////////////////////////////////////////////////////////
-    //! The class represents the FBX file structure and operations on that files.
-    //!
-    //! FBX file has a tree based structure.
-    //! Top structure of the FBX file consists of the following sections:
-    //!   - FBXHeaderExtension (mandatory) - metadata of file.
-    //!   - GlobalSettings (mandatory) - general data properties.
-    //!   - Documents (optional) - ?
-    //!   - References (optional) - ?
-    //!   - Definitions (optional) - ?
-    //!   - Objects (optional) - static data storage like objects, geometries, textures and materials.
-    //!   - Connections (optional) - defines connections between data defined in Objects.
-    //!                              For example which object uses specific material.
-    //!   - Takes (optional) - animations definitions.
-    //!
-    //! Additional documentation about FBX structure:
-    //! https://web.archive.org/web/20160605023014/https://wiki.blender.org/index.php/User:Mont29/Foundation/FBX_File_Structure
-    //! https://banexdevblog.wordpress.com/2014/06/23/a-quick-tutorial-about-the-fbx-ascii-format/
-    //!
-    //! Example FBX file
-    //! https://www.ics.uci.edu/~djp3/classes/2014_03_ICS163/tasks/arMarker/Unity/arMarker/Assets/CactusPack/Meshes/Sprites/Rock_Medium_SPR.fbx
-    class Fbx
+    namespace Fbx
     {
-    public:
-        using Property = std::any;
-        using Properties = std::vector<Property>;
-        enum class FileType { Text, Binary };
-
-        //! A node in FBX file tree structure.
-        //! Each named node could contain children nodes (subnodes) and properties.
-        class Node
+        ////////////////////////////////////////////////////////////////////////////////////////////////////
+        //! The class represents the FBX file structure and operations on that files.
+        //!
+        //! FBX file has a tree based structure.
+        //! Top structure of the FBX file consists of the following sections:
+        //!   - FBXHeaderExtension (mandatory) - metadata of file.
+        //!   - GlobalSettings (mandatory) - general data properties.
+        //!   - Documents (optional) - ?
+        //!   - References (optional) - ?
+        //!   - Definitions (optional) - ?
+        //!   - Objects (optional) - static data storage like objects, geometries, textures and materials.
+        //!   - Connections (optional) - defines connections between data defined in Objects.
+        //!                              For example which object uses specific material.
+        //!   - Takes (optional) - animations definitions.
+        //!
+        //! Additional documentation about FBX structure:
+        //! https://web.archive.org/web/20160605023014/https://wiki.blender.org/index.php/User:Mont29/Foundation/FBX_File_Structure
+        //! https://banexdevblog.wordpress.com/2014/06/23/a-quick-tutorial-about-the-fbx-ascii-format/
+        //!
+        //! Example FBX file
+        //! https://www.ics.uci.edu/~djp3/classes/2014_03_ICS163/tasks/arMarker/Unity/arMarker/Assets/CactusPack/Meshes/Sprites/Rock_Medium_SPR.fbx
+        class Fbx
         {
-            public:
-                Node(const std::string & name, const Properties & properties = {});
+        public:
+            enum class FileType { Text, Binary };
 
-                std::string GetName() const;
-                std::vector<Node> GetChildren() const;
-                Properties GetProperties() const;
-                bool HasChildren() const;
-                bool HasProperties() const;
+            //! Save the current FBX structure to file.
+            //! Note: only ASCII version is supported
+            void SaveToFile(const std::string & filePath, FileType type = FileType::Text);
 
-                void AddProperty(const Property & property);
-                void AddChildNode(const Node & child);
-                void AddChildNode(const std::string & name, const Property & property);
-                void AddChildNode(const Node && child);
+            //! Return the string with ASCII version of current FBX structure.
+            std::string GetFbxString();
 
-                std::string ToString(int nodeDepth = 0)  const;
+        private:
+            //! Generate the fbx file structure.
+            void GenerateFbxStructure();
 
-            private:
-                std::string m_name;
-                std::vector<Node> m_children;
-                Properties m_properties;
+            // Default FBX file header
+            Node GetFbxHeaderExtension() const;
+            Node GetTimeStamp() const;
+            Node GetSceneInfo() const;
+            Node GetMetaData() const;
+
+            // Default global settings
+            Node GetGlobalSettings() const;
+
+            Node GetDocuments() const;
+            Node GetDefinitions() const;
+
+            // Objects creation
+            Node GetObjects() const;
+            Node CreateGeometryCube(double size = 1.0) const;
+
+            Node GetConnections() const;
+
+            std::vector<Node> m_basicNodes;
+            bool m_nodesUpdated = false;
         };
-
-        //! Save the current FBX structure to file.
-        //! Note: only ASCII version is supported
-        void SaveToFile(const std::string & filePath, FileType type = FileType::Text);
-
-        //! Return the string with ASCII version of current FBX structure.
-        std::string GetFbxString();
-
-    private:
-        //! Generate the fbx file structure.
-        void GenerateFbxStructure();
-
-        // Default FBX file header
-        Node GetFbxHeaderExtension() const;
-        Node GetTimeStamp() const;
-        Node GetSceneInfo() const;
-        Node GetMetaData() const;
-
-        // Default global settings
-        Node GetGlobalSettings() const;
-
-        Node GetDocuments() const;
-        Node GetDefinitions() const;
-        // Objects creation
-        Node GetObjects() const;
-        Node CreateGeometryCube(double size = 1.0) const;
-
-        Node GetConnections() const;
-
-        std::vector<Node> basicNodes;
-        bool nodesUpdated = false;
-    };
-
+    } // namespace Fbx
 } // namespace ROS2
